+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @happyvertical/logger
 
+## 0.89.3
+
+### Patch Changes
+
+- @happyvertical/utils@0.89.3
+
+## 0.89.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/utils@0.89.2
+
 ## 0.89.1
 
 ### Patch Changes