+++ packages/repos/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @happyvertical/repos
 
+## 0.89.3
+
+### Patch Changes
+
+- @happyvertical/graphql@0.89.3
+
+## 0.89.2
+
+### Patch Changes
+
+- @happyvertical/graphql@0.89.2
+
 ## 0.89.1
 
 ### Patch Changes